pyproject.toml
[project]
name = "backendforphong"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "alembic>=1.16.4",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
]
